hacks: support symbolic links in .zip files
Add support for re-creating symbolic links when unpacking .zip files,
copying the complete implementation of ZipFile._extract_member() and
adding the needed modifications. As a safeguard to avoid following
symlinks when re-unpacking the same file, which could result in wrong
output, always unlink each destination file before creating it.

This is needed for newer versions of the Android NDK which e.g. link
"llvm-ranlib" to "llvm-ar" and a few others.