android: Remove __ANDROID_API__ definition from config
Simplify the android.cbc configuration file by removing the manual
definition of the __ANDROID_API__ macro. This is nowadays defined
by Clang automatically, and trying to redefine it from the command
line results in compiler warning spam in compilation logs. Removing
the manual definition is not only correct, it makes the logs more
readable.